Update and rename learn-to-race.tsx to intro-to-racing.tsx
