+++ src/containers/jp/class-signup/types/other/intro-to-racing.tsx
@@ -1,0 +1,17 @@
+import * as React from "react";
+import { ClassType } from "../../class-description";
+
+
+const data: ClassType = {
+	typeId: 1401,
+	typeName: "Intro to Racing",
+	prereq: "420 Green",
+	sessionCt: 5,
+	sessionLength: 2.5,
+	classSize: 14,
+	description: (<React.Fragment>
+		Intro to Racing is a dynamic one-week program designed to impart essential racing skills. Our comprehensive curriculum covers everything from racing rules and advanced sailing techniques such as roll tacking and gybing to spinnakers, racing starts, wind shifts, and ladder rungs. As participants progress through the course, they will develop their upwind sailing abilities, fostering confidence and competence on the water.
+	</React.Fragment>)
+}
+
+export default data;
