--- src/containers/jp/class-signup/types/other/learn-to-race.tsx
@@ -1,17 +1,0 @@
-import * as React from "react";
-import { ClassType } from "../../class-description";
-
-
-const data: ClassType = {
-	typeId: 1401,
-	typeName: "Learn to Race",
-	prereq: "Mercury Yellow Rating",
-	sessionCt: 3,
-	sessionLength: 3,
-	classSize: 16,
-	description: (<React.Fragment>
-		Sailboat racing can teach sportsmanship, teamwork, and responsibility. In addition to inspiring many with a lifelong passion for sailing, racing can vastly improve sailing ability. This Learn-to-Race class is three days long and will take place in our Cape Cod Mercuries. Sailors must have at least a Mercury Yellow rating, but more advanced racers and sailors are welcome and encouraged to join. For sailors who are interested in having fun and improving their skills, this Learn-to-Race class is a great place to start.
-	</React.Fragment>)
-}
-
-export default data;
